requirements.txt
# requirements.txt para Strava Telegram Bot
python-telegram-bot
stravalib
